small_library/requirements.txt
Django==3.2.8
django-redis==5.0.0
djangorestframework==3.12.4
redis==3.5.3
celery==5.1.2
graphene-django==2.15.0
